pyproject.toml
[project]
name = "math-function-tool-agent-assignment-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.9",
    "python-decouple>=3.8",
]
